be/aa-12751. QA Fix 2
