+++ packages/cli/src/WorkflowRunner.ts
@@ -59,6 +59,7 @@ import { eventBus } from './eventbus';
 import { recoverExecutionDataFromEventLogMessages } from './eventbus/MessageEventBus/recoverEvents';
 import { Container } from 'typedi';
 import { InternalHooks } from './InternalHooks';
+import { logIncidentFromWorkflowExecute } from './lib/incidentLogger';
 
 export class WorkflowRunner {
 	activeExecutions: ActiveExecutions;
@@ -340,10 +341,10 @@ export class WorkflowRunner {
 					data.executionData,
 				);
 				workflowExecution = workflowExecute.processRunExecutionData(workflow);
-				console.log(
-					'processRunExecutionData in WorkflowRunner:',
-					(await workflowExecution).data.resultData,
-				);
+				workflowExecution.then(async (data) => {
+					await logIncidentFromWorkflowExecute(data, workflow);
+					return data;
+				});
 			} else if (
 				data.runData === undefined ||
 				data.startNodes === undefined ||
