+++ packages/cli/src/WorkflowRunnerProcess.ts
@@ -58,19 +58,7 @@ import { PermissionChecker } from '@/UserManagement/PermissionChecker';
 import { License } from '@/License';
 import { InternalHooks } from '@/InternalHooks';
 import { PostHogClient } from '@/posthog';
-import { createIncidentLog, generateCreateCustomTicketSubjectFn } from './lib/incidentLogger';
-import { setupReusablesAndRoutes } from './reusables';
-
-const getIncidentHandlingFN = (() => {
-	let createIncidentLogFN: typeof createIncidentLog;
-	return async () => {
-		if (createIncidentLogFN) return { createLog: createIncidentLogFN };
-		const app = express();
-		await setupReusablesAndRoutes(app);
-		createIncidentLogFN = createIncidentLog;
-		return { createLog: createIncidentLogFN };
-	};
-})();
+import { logIncidentFromWorkflowExecute } from './lib/incidentLogger';
 
 class WorkflowRunnerProcess {
 	data: IWorkflowExecutionDataProcessWithExecution | undefined;
@@ -306,40 +294,7 @@ class WorkflowRunnerProcess {
 
 			const data = await this.workflowExecute.processRunExecutionData(this.workflow);
 			console.log('processRunExecutionData in WorkflowRunnerProcess 2:', data.data.resultData);
-			if (data.data.resultData.error) {
-				const { runData, error } = data.data.resultData;
-				// eslint-disable-next-line @typescript-eslint/prefer-nullish-coalescing
-				const nodeStack = data.data.executionData?.nodeExecutionStack || [];
-				const errorNodeName = Object.keys(runData).find((nodeName) => {
-					return runData[nodeName].some((data) => {
-						return data.executionStatus === 'error' && data.error;
-					});
-				});
-				const errorNode = nodeStack.find(({ node }) => {
-					return node.name === errorNodeName;
-				});
-
-				const { createLog } = await getIncidentHandlingFN();
-				console.log('Attemping to create log...');
-				try {
-					await createLog(
-						{
-							errorMessage: error?.message,
-							incidentTime: new Date(),
-						},
-						{
-							workflow_id: this.workflow?.id,
-							error_node_id: errorNode?.node.id,
-							error_node_type: errorNode?.node.type,
-							error_message: error?.message,
-						},
-						generateCreateCustomTicketSubjectFn(Object.values(this.workflow?.nodes)),
-					);
-				} catch (error) {
-					console.log('ERROR when attemping to create incident log:', error);
-					throw error;
-				}
-			}
+			await logIncidentFromWorkflowExecute(data, this.workflow, true);
 
 			return data;
 		}
